more clear and consistent naming of types
Use singular for the most part. For the most part the tables contain many of something (`pagelinks`) while the rows contain just one (`PageLink`), so the plural naming convention for some of the row types (`PageLinks`) was misleading. `SiteStats` is the only one for which plural is accurate, because it contains multiple figures.

Expand abbreviated `ChangeTagDef` to `ChangeTagDefinition`