add target_id field to TemplateLink
MediaWiki 1.38 added the `tl_target_id` field to
the `templatelinks` table. This field is now present in
`templatelinks.sql` in recent English Wiktionary dumps.
`templatelinks.sql` produced by earlier MediaWiki
versions will no longer be able to be deserialized after
this change.

The `linktarget` table, to which this field points, is not yet
provided in the dump, and no type has been added for it.
So the `LinkTargetId` is an orphan here.