use Title::new_unchecked that doesn't clone String
`Title::new_unchecked` formerly unconditionally
replaced underscore with space in the `dbkey`, which
requires cloning the `String`, and was unnecessary
when the `dbkey` comes from `page.sql`. mwtitle 0.2.0
(in alpha) pushes cloning to the caller of the function,
so you can avoid memory allocation. This change
reduces allocations in several examples.