show short error message in template_redirects example
Avoid printing the debug representation of the contents of
page.sql when it can't be parsed.

The printing was causing an attempt to allocate about 3 or 6 gigabytes in the grid job engine on the Wikimedia Toolforge
servers. Way too much and not helpful at all. I based the error
printing code from test.rs, which shows where the
problem actually occurred. In this case there was a missing field
in the SQL, from the perspective of the Page parser.